pyproject.toml
[project]
name = "datamind-ai"
version = "2.1.0"
description = "Agent d'analyse de donnees intelligent propulse par LLM"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.6.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "plotly>=5.18.0",
    "openai>=1.30.0",
    "python-dotenv>=1.0.0",
    "scipy>=1.11.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "llama-cpp-python>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_phase*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["agent", "backend", "utils", "prompts"]

[tool.setuptools.packages.find]
include = ["agent*", "backend*", "utils*", "prompts*"]
exclude = ["frontend*", "tests*", "data*"]
